Fix mouse push failing; rename pre/post-command hooks
The root cause is that sit-for returns whether it actually waited
for as long it was asked for. In the advice :around, this return
value was being thrown away. Correctly bind and return this value
in the advice `explain--wrap-sit-for`.

Also, noticed that other packages were naming their pre and post
command hooks as `xyz-pre-command-hook`, which makes sense. Do
the same thing.