Fix printing problems caused by unescaped %
`message` always treats the first argument as a format string,
this means that if the string contained %, then it would expect
an argument after. If a command-set contained a lambda or
byte compiled function, the closures or references for that
might contain a `%`. Fix this by regex escaping all `%` into
`%%`.