Create 3 levels of alerting, and unify alert and profile messaging. (#32)
Implement 3 levels of alerting. Never implies none.
Normal collects alerts and after a certain number of them, alerts
the user with a count, but only after enough time has passed.
It is fairly unobtrusive, and the defaults should mean at most,
one sees it every 15 minutes if their emacs is extremely slow.

Developer implements something similar to the earlier behavior,
however alerting about slow commands and profiling events are
unified. Messages are only written to the minibuffer if it is not the
active buffer.

Additionally, regropus and renames defcustoms to be more manageable
and hides the developer / advanced options in new subgroups.

Also, all alerting now uses the measured-command hook introduced
in #23.

Fixes #13.