Fix fringe and margin header and table width display bugs (#36)
1. Fringe and margin WAS correctly accounted for in table width,
but was not accounted for when displaying the header. Fix this
using a propertized spacing, because it's a pixel size, not a
character size.

2. The "optimized" code to handle resizes of windows didn't work
if only one window resized itself down, because it was trying
to be smart and only resize upwards. Remove all that and just
cycle over all the windows.