Fix display crash OOB when items exist on first render of `top' (#43)
The number of columns used to be calculated from the number of
display-column-widths. On first render, this is nil, because
needs-resize is t, and nothing has yet been calculated. This however
still worked so long the buffer rendered once before adding new
entires. If entries existed on first render, the column count would be
0.

Fix by calculating column count off of header-widths, which is always
set correctly.