Add integration tests
Full integration tests that run emacs in tmux so we can press keys
and fully simulate a user. Use the event logging stream to connect
to the driver, which is also written in elisp. Drive the driver
via Make. Add test cases for most existing (fixed) bugs that are
not covered at unit test layer.