Return the original wrapped sentinel/filter function when requested.
When we wrap a process filter or sentinel, we end up putting in a
lambda into the actual process. If later, `process-filter` is called
to get the filter out, eq comparisons that used to work will fail.

Store the original callbacks into the process object so we can
hijack `process-filter` to return the original.

Fixes #46. Originally found from `magit`, in `with-editor`,
in `with-editor-set-process-filter`.